separate order of grid items from alignment
When the order of grid items is reversed, the desired behaviour is not
necessarily to align the entire grid to the right. It can be seen when
grid items widths don't add up to 1 (for example, they add up to three
quarters). Better to separate concepts of order and alignment.

 - removed `text-align:right` from `grid--rev` and added
`text-align:left`
 - added `grid--right` modifier to manage right aligment of grid;